Move `version` command to a separate module
