+++ textx/cli/__init__.py
@@ -10,16 +10,6 @@ def textx(ctx, debug):
     ctx.obj = {'debug': debug}
 
 
-def version(textx):
-    @textx.command()
-    def version():
-        """
-        Print version info.
-        """
-        import textx
-        click.echo('textX {}'.format(textx.__version__))
-
-
 def register_textx_subcommands():
     """
     Find and use all textx sub-commands registered through the extension point.
